respond to feedback:
- more opinionated shipping context
- more explicit input error message detection in specs
